requirements.txt
streamlit==0.79.0
seaborn==0.11.1
matplotlib==3.3.4
pandas==0.25.3
numpy==1.18.1
Pillow==8.1.2
scikit_learn==0.24.1
